refactor(runners): move video.py out of evals/ to runners/video.py
Per review: evals/ holds evaluation environments/backends (robosuite, isaac_lab,
cosmos_reason); the rollout-video helper is a cross-runner *utility* that
produces an artifact, not an environment. It now sits beside the other runner
infrastructure (subprocess.py, base.py). Only the two imports updated.
