+++ tests/unit/test_rollout_video.py
@@ -16,7 +16,7 @@ from odyssey.engine.lifecycle import TaskStatus
 from odyssey.engine.records import MissionRun, TaskRun
 from odyssey.runners.base import TaskContext
 from odyssey.runners.evals.robosuite import RobosuiteRunner
-from odyssey.runners.evals.video import save_rollout_video, to_uint8_frame
+from odyssey.runners.video import save_rollout_video, to_uint8_frame
 from odyssey.spec import (
     AgentRole,
     AgentSpec,
